scicite-back-main/requirements.txt
Django==4.2.1
django_filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf_yasg==1.21.7
python-decouple==3.8
Unidecode==1.3.6
psycopg2-binary==2.9.8
celery==5.3.4
django-composite-field==1.1.0
django_cors_headers==4.3.1
django_celery_beat==2.5.0
django-celery-results==2.5.1
pandas==2.1.4
openpyxl==3.1.2
redis==5.0.1
gunicorn==23.0.0
django-modeltranslation